FIX: Compiler warnings fixed. ADD: ZEFlags class added. REM: ZEToolSDK types removed. Foundation types will be used. IMP: Improvements on ZEToolSDK coding style and source files.
svn path=/trunk/; revision=630

+++ Source/ZECore/ZECore.cpp
@@ -282,7 +282,7 @@ float ZECore::GetFrameTime()
 float ZECore::GetRuningTime()
 {
 	QueryPerformanceCounter(&PerformanceCount);
-	return (PerformanceCount.QuadPart - StartPerformanceCount.QuadPart) / PerformanceCounterFreq.QuadPart;
+	return (float)((PerformanceCount.QuadPart - StartPerformanceCount.QuadPart) / PerformanceCounterFreq.QuadPart);
 }
 
 
@@ -617,7 +617,7 @@ void ZECore::MainLoop()
 	{
 		QueryPerformanceCounter(&PerformanceCount);
 
-		FrameTime = (double)(PerformanceCount.QuadPart - OldPerformanceCount.QuadPart) / (double)PerformanceCounterFreq.QuadPart;
+		FrameTime = (float)((PerformanceCount.QuadPart - OldPerformanceCount.QuadPart) / PerformanceCounterFreq.QuadPart);
 		OldPerformanceCount = PerformanceCount;
 	}
